// Visitor Manager UI/src/app/model/visitor.ts
export class Visitor {
    Id: number;
    VisitorName: string;
    VisitorType: string;
    TotalVisitor: number;
    WhomToVisit: string;
    VhicleType: string;
    VhicleNumber: string;
    TimeIn: string;
    TimeOut: string;
    Date: string;
    AddDate: string;
    Image: string;
    StaffId: string;
    Mobile: string;
    base64: string;
    Status: string;
    AdminId: string;
    OTP: number;
    Purpose: string;
    GatekeeperId: number;
}